Сделал User Service с логикой

+++ src/main/java/ru/skypro/homework/repo/AdsCommentRepo.java
@@ -1,8 +1,7 @@
-/*
+
 package ru.skypro.homework.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdsCommentRepo extends JpaRepository<AdsCommentRepo, Long> {
+public interface AdsCommentRepo extends JpaRepository<AdsCommentRepo, Integer> {
 }
-*/
