+++ src/main/java/ru/skypro/homework/repo/AdsRepo.java
@@ -1,7 +1,7 @@
 package ru.skypro.homework.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.skypro.homework.entity.Ads;
+import ru.skypro.homework.entities.Ads;
 
-public interface AdsRepo extends JpaRepository<Ads, Long> {
+public interface AdsRepo extends JpaRepository<Ads, Integer> {
 }
